refactor: enrichPerformance 팩터리 메서드 이동
